Added Java Project to Repository
In Java, an abstract class is a class that cannot be instantiated on its own and is typically used as a base for other classes. It may contain both abstract and concrete methods. Abstract methods are declared without implementation and must be implemented by subclasses. Abstract classes can also contain regular methods with implementations. Abstract classes serve as a blueprint for other classes to inherit from, providing common behavior and structure while allowing subclasses to customize specific functionality.